Remove customised sentry error logger
This is because it duplicates sentry errors. It seemed like there was a previous issue of errors not sending to sentry correctly, however, atm this is not the case.

Use sentry-sidekiq gem to include the sentry logger.

This also means the sentry error logger doesn't send PII (personal identifiable information) to sentry. The error is now captured by sentry and santized before being sent.