AP-2082 Add initial DWP override functionality
Solicitors need to be able to override a DWP benefit check result.
This commit amends the check_benefits page so that after a negative
result the solicitor is given the option of overriding it. For now this
simply leads to a holding page.

- amend the check_benefits_controller
- refactor the check_benefits views
- add necessary translations and flows

This functionality is initially hidden behind a feature flag.