Bugfix: Re-add old CFE Result models
AP-2035 removed v1 and v2 of the CFE Result model. This can cause
issues as the cfe_results table uses single table inheritance. When a
record that uses CFE::V1::Result or CFE::V2::Result is read or updated
it will throw an error if the model is not present.

There is minimal risk of this happening but it seems safer to leave
retain the old models.