Identify renamed tables in migration_hook
The migration_hook used by the 'Check anonymise rules' GitHub action
does not identify cases where a table has been renamed. It only alerts
if there is a difference in the number of tables between the rules file
and the schema.

This commit amends the script to identify any difference in table names
by finding the difference between the two arrays and failing with an
appropriate message.