Fix strange capybara bug for long page title
When the page title is long and the column_width of the page layout is not 'full'
it will cause a scroll on the page. However, this causes capybara to fail to click 'Save and continue' submit button.

The solution was to add a step to scroll down the page