Add migration files for legal framework related tables
Add migration files to add a submissions table, submission histories and merits task list table for the legal framework

The merits task list table contains the serialized data and a reference to the legal aid application. The serialized data
is the legal framework api call response but in yaml format.