Move and update controllers relating to ChancesOfSuccess model
Move the chances_of_success controller and success_prospects under the proceeding_merits_task folder
as they are dependant on each other.

Leave the proceedings_before_the_courts_controller under the app/controllers/providers folder because it is not
a proceeding_merits_task at the moment (could change in the future).

Update these controllers to take in an application_proceeding_type_id and update/create its chances_of_success association

Add a legal_aid_application method on these controllers because it is used by the ProviderBaseController