Update the LegalAidApplication model for chances_of_success
Remove the chances_of_success association with LegalAidApplication because it's now associated
with the ApplicationProceedingType model

Delegate the chances_of_success method call to lead_application_proceeding_type. This is because of files like the
ApplicationDetailCsvLine delegation issues.

Add a lead_application_proceeding_type method

Add a merits_complete! method which updates the merits_submitted_at attribute with the current date.