Update the provider merits flow for chances_of_success
Update the flow for success_prospects, chances_of_success and proceedings_before_the_court
to use the lead application proceeding type instead of the legal aid application.

This meets the new multiple proceedings feature.

Add a nocov for proceedings before the court as that page is not being used. Leaving it as TechnicalDebt for now.