Update the mis report to get chances of success
Get the chances of success from the lead application proceeding_type instead of the legal aid application

Also  has the merits_submitted_at under the chances_of_success, which isn't accurate but it can be done in a following ticket e.g. to update the excel sheet and test it.