AP-2191: Add delegated functions page for the multiple proceedings journey
This commit adds a new page to the provider flow to accommodate the
ability to add delegated function dates to each application proceeding
type. Initially a db migration has been added that adds a new 'name'
column to the proceeding types table and populates it with attribute
names for each type. This is to reduce the need to parse proceeding
meaning names to something consistent that works for all the checkbox
and date field attribute names and is Rails friendly. References to
whether delegated functions exist and their respective date has started
to be moved to application proceeding types. Additions have been made to
the application proceeding types model to allow any proceeding type to
inform:

1) Have delegated functions have been used
2) What application proceeding type has the earliest associated delegated functions
3) What the earliest delegated functions date is

Updates to html pages to reveal the necessary info when multiple
proceedings is turned on has been made whilst preserving the original
single proceedings journey. Notes have also been added to inform others
as to what will need to be refactored as the rest of the multiple
proceeding work is merged.