Update the CFE base result to work with version 4 results
Update the base result to work with version 4 cfe api results
and add a version_4? method to be used when giving a dynamic result for v3 or v4 results

Add an ampersand to the .to_d method because this method turns a nil to 0.0. This is because when looking
for a key in the cfe result, if it doesn't find it, the parsing returns nil which can be falsey translated to 0.0 with to_d