AP-2359: Reduce Dashboard::UpdaterJob calls to geckoboard
Previously we would update geckoboard after each step of the CCMS `dance`
That could mean 20-30 calls in a few seconds and would occasionally trip
the geckoboard rate limits

This PR attempts to reduce that noise bu sending a ccms_submission_saved job
when a CCMS submission is initialised and then updating again when it concludes
either completed or failed.  It also updates the Applications call to send
whether the job completes successfully or not.  This should clear the
CCMS error widget when manually pushing a CCMS job though