Seperate flickering test for multiple used delegated functions
Seperate out the test for multiple used delegated functions as this is a flickering test.

This will enable us to narrow down the cause of the flickering test if it happens again.

An assumption could be the application proceeding types are returning in different order because of the factory. Just an assumption.