Update class name for removing the govuk summary list border
Update html class '.govuk-summary-list--no-border' to class '.govuk-summary-list__row--no-border'

This is done for row by row basis. This is because it was not working when printing the check merits answers
page. The print page would still show the border. Now with the __row--no-border it no longer shows as expected.
This is because the class was being added to a summary list row and thus required a row by row basis to remove the
border e.g.  'govuk-summary-list__row--no-border'