BugFix: Address BigDecimal rounding issues
There was another flicker when the truelayer MOCK_DATA generated
a value of `823332.18`, the test found `823332.1800000001`
previous experience shows this would have been fine at
`823332.19`

This solution converts the float to a string and passes it to
BigDecimal.  Passing the float directly requires the precison
to be sent, which requires converting to a string!

I've backported this solution to a previous bugfix as it's easier
to read and means the code is more consistent