Allow for existing attachments with no original filenames
Some attachments in our current databse does not have original filenames values in their records.
This caused a bug because the attachment helper used on the check merits answers page was expecting something.

using a generic attachment_type name e.g. 'statement_of_case' as the name for the file when non exists is a temporary fix.

This should not be the case going forward because all attachments will now have original_filenames attached.