Update ccms submission worker to retry with sidekiq
Allow sidekiq to do its error handling job and to retry workers following its automatic job retry.

Add a middleware to sidekiq to gather the retry count of a worker and to place it in an attribute reader for the
ccms/submission_process_worker

Add a retry of 10 times which should last around 4 hours of retrying before failing.

Sidekiq is already configured to raise errors to Sentry and therefore the retries will be raising the errors if it fails.

Add test for worker