Refactor base report and reports tests to be more accurate
Refactor the creation of ccms_submission with #process! method so that it's easier to test.

Create a new method on the described_class called process_ccms_submission which means it doesn't fail because
testing legal_aid_application.create_ccms_submission and legal_aid_application.ccms_submission.process! in the same
test block results in the ccms_submission being nil when trying to process! as if it needs to wait.