Fix: Change pause_submission event handling
This was originally implemented at haste due to the
short notice of CCMS shutdown.  It was not clear
when developing at pace that the first call to the
CCMS APIs were made in the ReportsCreatorWorker.

Therefore when we 'paused' teh submissions the reports
creator was failing to create the documents as there
was no way to fulfill the `ensure_case_ccms_reference_exists`
by requesting a `case_ccms_reference`

This PR moves the pause step forward to after check_merits
and removes the check on whether the step should run in the
controller