AP-2472: Add abandoned state to CCMS::Submission
This allows us to manually set a submission to abandoned
and exclude it from searches.  This should prevent a
repeat of the situtaion where a submission failed and
it was re-initaliased from scratch but the original
submission sat in the DB as failed for a year!

Callable with `submission.abandon!`