AP-2448 Add translation for failed address lookup
If a postcode lookup fails due to a connection failure (eg if wifi
is turned off), the user should see the manual address entry screen with
an appropriate message however instead  an error is presented to the
user.

This commit adds the correct wording to the config/locales/en/forms.yml
locale. It also corrects the wording which is displayed when an
unsuccessful response is received from the OS API, as it is currently
misleading.