AP-2501 Incorrect legend on /check_client_details
A change to the DFE GOVUK Formbuilder has resulted in an incorrect
legend being displayed for the radio buttons on /check_client_details.

This is because "legend: { text: '' }" now falls back to the default
value ('Check_client_details') instead of not displaying a legend.

This commit replaces it with "legend: nil" so that no legend is
disaplyed.