AP-1946 Fix failing check provider answers specs
Delegated functions tests in check_provider_answers_spec are failing
after the removal of the multiple proceedings flag.

These tests were never made compatible with multiple proceedings and
are failing due to a change in the structure of the page. This commit
fixes the spec so they now look in the correct location for the
delegated functions data. It also fixes an issue with the
proceedings_details partial which was preventing this working correctly.