AP-1946 Fix failing means report creator tests
Tests in means_report_creator_spec are failing following the removal of
the multiple proceedings feature flag.

Test data in spec/factories/cfe_results/v4/mock_results.rb hardcodes the
proceeding types needed for these tests as DA002 and DA006. This factory
used to create the application defaults to DA001, so there is always a
mismatch leading to failures.

This commit fixes that by explicitly specifying DA002 and DA006 as the
proceeding types on the application.