AP-1946 Fix failing check merits answers tests
Tests in check_merits_answers_spec are failing for two reasons:

1) Change links for prospects of success now exist for each proceeding
and contain the uuid of the proceeding in the url. The existing
ChangeLinkMatcher does not cater for this so the tests fail. This
commit amends that class so that if a success_likely link is present it
is matched using the uuid of the proceeding.

2) The wording of the Statement of Case question has been changed from
'Provide a statement of case' to just 'Statement of case', however the
translation file which the test uses has not been updated. This commit
fixes the wording.