Fix flashing feature test
A 'check your answers' feature test is occasionally failing because the
application used is randomly allocated a proceeding type. If this
proceeding type is a Section 8 proceeding the test will blow up, as
there must always be a domestic abuse proceeding on an application.

This commit forces the proceeding type for this test to always be a
Domestic Abuse proceeding.