AP-2464 Fix failing tests
Adding callbacks to the ApplicationProceedingType model cause existing
tests to break, as not all tests set up proceeding types with scope
limitations.

The options to fix this appear to be (1) to change all failing tests to
create scope limitations, which is inefficient and will require a large
number of changes to factories, or (2) to disable those callbacks for
all tests and explicitly enable them where relevant.

This commit does (2).