AP-2455 Update CircleCI Slack Orb to 4.4.4
* Updates CircleCI Slack Orb to 4.4.4

* Replaces slack/status command used to report on failures of the
overnight full test run with a customised slack/notify command. The
customisation is necessary to include instructions in the notification
on manually running the CCMS tests to resolve the failure.

* Replaces hold_production_notification with a customised slack/on-hold
command. Customisation isn't strictly necessary but is used to make the
notification look similar to the existing format and get rid of the
oversized CircleCI logo.