Fix skipped feature test
One feature test was being partially skipped because one step included
text 'You'll need to tell us if your client' in single quotes. The
apostrophe caused the step not to be recognised and all subsequent steps
to be skipped.

This commit switches double quotes to allow the whole test to run, and
adds an additional missing step so that it passes.