AP-2633 Update UAT values/deploy script for live
Create new values-live-uat.yaml to be used when deploying the app to the
new live cluster. This initially has aws-weight set to 0; this will be
increased when we are satisfied that it is stable.

Create new uat_deploy_to_live script to deploy the app the the new live
cluster. Sets the identifier as -green to allow traffic to be balanced
across both clusters.