Fix live deployments for long branch names
On branches with long names, deployments to the live cluster fail
because the annotation we set in uat_deploy_to_live does not match the
truncated ingress name expected by cloud platform.

This commit updates the uat_deploy_to_live so that it also truncates the
ingress name, which allows deployments to succeed.

It also updates uat_deploy for consistency between live and live-1.