AP-2565 Call HMRC from non-passported DWP page
Calls to HMRC are now made after a provider confirms that their client
is not passported.

This commit adds the call to the confirm_dwp_non_passported_applications_controller
protected by both a feature flag and user-level permissions.

Includes a small change to the legal_aid_applications factory to fix
a failing test.