AP-2719: BUGFIX - Substantive scope limitations on DF apps
Substantive scope limitations are incorrectly appearing on emergency
applications in CCMS.

This is because the generate_scope_limitations method in the
CaseAddRequestor class sets the DelegatedFunctionsApply attribute as
'proceeding.used_delegated_functions?' for both the substantive and
delegated functions scope limitations. This means that for a DF
application, both scope limitations are flagged as
DelegatedFunctionsApply.

This commit changes that method so that the substantive scope limitation
always has a DelegatedFunctionsApply attribute of false, and the
delegated functions scope limitations always has a
DelegatedFunctionsApply attribute of true.