Fix flaky chances of success tests
Tests in spec/models/proceeding_merits_task/chances_of_success_spec.rb
are occasionally failing due to the way application_proceeding_types
and proceedings are set up.

By creating the proceeding after the creation of the application and
application_proceeding_types, and by creating different numbers of
application_proceeding_types and proceedings, it is possible for the
proceeding_case_id to be non-unique.

This commit changes the set up so that identical
application_proceeding_types and proceedings are created simultaneously.
This should prevent the unique constraint being violated.