AP-2570 Include 'No HMRC data' page in flow
Amends the provider flow so that new no_employment_income page is
displayed if no employment income data is received from HMRC, and the
employment_income page is displayed if such data does exist.

Adds methods to the LegalAidApplication and HMRC::Response models to
make identification of this data easier.