AP-2814 Remove with_substantive_scope_limitation trait
The with_substantive_scope_limitation trait on the legal_aid_applications
factory is no longer needed following the move to using LFA for proceeding
data. It is marked as deprecated, to be replaced by with_proceeding_types
which has been removed in a previous commit.

This commit deletes the with_substantive_scope_limitation and amends any
tests still using it to ensure they continue to pass.