AP-2813 Remove proceeding_types from CFE Result v4
Updates v4 of the CFE Result model to replace a reference to deprecated
ProceedingType model with Proceeding. Also updates tests, factories and
fixtures.

Leaves some references to 'proceeding_types' in code as these derive from
CFE and a) it's out of scope for this ticket to change CFE, and b) it
makes sense for methods/variables to continue to be named after what we
receive from CFE.