Fix mock HMRC data generation
The HMRC::MockInterfaceResponseService does not set the submission_id on
the HMRC::Response record. This causes the process to fail as the
submission_id is used to subsequently update the Response with the
correct json payload.

This commit changes the MockInterfaceResponseService so that it
generates a submission_id UUID, and updates the Response record with it,
in its initializer.