AP-2851 Update UAT ingress host
We need to remove reference to the legacy live-1 cluster from ingress
hosts. This will be done over several stages.

As part of the first stage this commit updates the UAT deployment script
to use both the old and new ingresses. The current UAT host is held as
an environment variable in CircleCI. This is different to our other
repos and basically just obfuscates the deployment process, so this
switches to defining it in the deployment script.