AP-2851 Update UAT ingress config
helm_deploy/apply-for-legal-aid/values-uat.yaml contains ingress
configuration which is not needed and causes problems when deploying
using multiple ingresses. This commit removes the following lines:

  hosts:
    - chart-example.local
  tls: []

and amends deploy/helm/templates/ingress.yaml to allow multiple hosts
to be set. This requires changes to the values-staging.yaml and
values-production.yaml to make them compatible.

Following these changes the Apply ingress configuration is consistent
with the configuration of the other three applications.