AP-2739 Refactor existing evidence validation tests
Enables a number of validation tests in the 'continue button' context of
/spec/requests/providers/uploaded_evidence_collection_spec and moves
them to the 'upload button' context.

This makes more sense as validation is done on upload of the file, not
when moving to the next page.

This will allow the 'continue button' context to be used for validating
business rules rather than file integrity.