AP-2739 Make required benefit evidence dynamic
On /uploaded_evidence_collection the bullet point that tells the
solicitor to upload evidence of benefits is hardcoded to display
Universal Credit. This should display the benefit received by the
applicant as entered on the DWP override page.

This updates the view and translation to allow the benefit to be
dynamically populated.