AP-2739 Refactor delete functionality
The delete button on the evidence upload page causes submission of
evidence to fail - rails always defaults to the last action on
the page, regardless of which is clicked. In this case it's the delete
action, so clicking Save and Continue actually deletes the files...

Instead, direct the Delete button to the patch action, check which
button was clicked, and route to destroy if necessary.