Ap-2739 Move evidence upload validation to model
Validation of business rules for uploaded evidence is currently done in
the submission_form, however this results in error messages being
displayed in the middle of the screen (above the submission form)
instead of at the top of the screen. This moves validation from the form
to the model, and use the base page template to render errors for the
model.