AP-2877 Fix column widths on Employment Income Review
When Tax or National Insurance values on the Employment Income Review
screen have more then five significant digits they exceed the column
width and wrap incorrectly.

This commit amends the widths of the table header columns to match the
widths of the table body columns which improves the amount of space
available for data.