AP-2866 Parse HMRC Responses with no employments
When making a CFE submission for an applicant that is employed but is
not found in the HMRC API, the parsing of the HMRC Response fails.

This is caused by a check in the CreateEmploymentsService which assumes
that if an applicant is employed then they will have HMRC data; this
will not always be the case.

This commit replaces the call to `legal_aid_application.employed?` in
the guard before processing the response with
`legal_aid_application.hmrc_employment_income?` which ensures that we
only try to parse employment data when the data actually exists.