AP-2590 Use all manual employment info in ManualReviewDetailer
Updates the ManualReviewDetailer service class to use the new
manually_entered_employment_information method so that the decision to
flag an application for review reflects any employment income data
entered by the provider.

This means that applications where no data is returned by HMRC and the
provider enters all the data manually will now be flagged for caseworker
review.