AP-2935 Add GOVUK Rubocop Gem
Replaces the generic rubocop gem with the GOVUK version to ensure we
meet cross-gov code standards and have consistency between our
applications.

Adds all offences to a (very long) .rubocop_todo.yml file so that they
can be addressed later.

Where we have previously overriden cops in .rubocop.yml, I have left the
configuration in if it targets specific file/s (as we know we have
unresolvable issues with some classes) but removed it if it was a
generic change, as we should now be aiming to conform to the GOVUK
standard.