Fix flaky rspec array comparison
In spec/models/legal_aid_application_spec.rb we test the output of
uploaded_evidence_by_category, which is a hash of arrays. The elements
of those arrays are not always populated in the same order, which
causes failing tests. We can't use the match_array method directly as
the arrays are nested within a hash.

This commit adds a deep_match method which allows for comparison of
nested data structures and updates the test to use it.