Minimal fix to use `net_housing_costs` instead of `rent_or_mortgage`
`net_housing_costs` deals with caps and other housing cost
edge cases, as well as - i believe from conversions with
other devs - being a month equivalent.

However this feels icky as it breaks the basic pattern of
having the values for moe in the `month_equivalent` response
from CFE.

Should this be handled in CFE instead and apply remain
unchanged?