Do not alert on use_case "two" responses
Quieten the sentry output from HMRC responses.

Every query to HMRC interface will result in two
response. a use_case one and a use_case two. The
use_case two are not currently valid or usable so
are always considered unpersistable for employment
purposes and so we do not need to alert on them.