Hardcode section 8 level of service value
The default "level of service" for a section 8 matter
in CCMS if Family help (higher), with a value of "1", not
as the current hadrcoded level-of-service, "Full representation"
with a value of "3".

This is the simplest and most immediate approach to sending a correct
default value for the section 8 matter type. Longer term we are looking
to store this data for the 6 level-of-services in LFA.