Do not validate on first or last name
Avoid mismatches resulting from unusual
names and spelling mistakes either by our
users or in HMRC records.

We want some level of validation so we can
be confident our request of HMRC results in
a response for the individual/applicant entered
in our system. We have decided as a team (CB/JS/SR)
that the unique nature of a nino coupled with a date
of birth should suffice to give us warning of inappropriate
responses from HMRC.