Use custom github action to delete UAT release
This is an alternative to using a local repository
github action, which could be copied from
`jsugarman/delete-uat-release-action`.

This could allow us to share with other apply
services - CFE, LFA, HMRCI.

If this is a better approach then we can move
the github action to an moj apply service specific
repo.