Use bundler-audit to audit gems
This adds the `bundler-audit` gem to automatically check installed gems in
Gemfile.lock for security advisories, and prompt us to upgrade to patched
versions as soon as possible.

This uses GitHub actions to run the check as part of CI, and it is triggered by
default pull_request events (opened, synchronized, reopened) against the main
branch. This seemed to be the trigger used across the application, so I have
adopted the same approach.