Remove `enable_employed_journey` setting
Before, this setting could be used to enable/disable the employed journey for
users so the feature could be incrementally deployed and released.

Now the feature has been turned on in all environments (and won't be turned
off), the setting can be removed.

This removes all references to the `enable_employed_journey` setting from the
code, although it leaves the database column in place.

A subsequent migration will be run to remove the ignored column once this change
has been successfully deployed to production.

[AP-2637](https://dsdmoj.atlassian.net/jira/software/c/projects/AP/boards/233?modal=detail&selectedIssue=AP-2637)