Remove `enable_employed_journey` column from settings
Before, in #4216, the `enable_employed_journey` setting was removed and the
associated column was ignored.

Now that change has been deployed successfully, we can safely remove the column
from the database table.

This removes the column from the settings table and removes the
`.ignored_columns` method defined in the Setting model.