Add capybara support to integration tests
This project leans heavily on cucumber for feature tests, and we write **no**
RSpec system specs at all.

In my opinion this is quite unusual, and a side effect of this is that we use
RSpec request specs (which are a wrapper around
ActionDispatch::IntegrationTestCase) a lot to assert against the rendered
content.

This isn't necessarily a bad thing, since Rails integration tests can offer lots
of benefits over system tests (e.g speed, greater control over HTTP verbs and
this session etc).

While I think this combination makes it harder for us to do
accessibility-focussed test-driven development (Capybara and RSpec system specs
are excellent for this), that conversation is one for another day.

This proposed change makes use of an abstraction discussed at length in [this
blog post](https://thoughtbot.com/blog/integration-testing-with-capybara) in
order to bring Capybara into our integration tests.

I've made changes to `spec/requests/errors_spec.rb` to highlight some of the
improvements this change could bring us.

I'm not suggesting we re-write all our specs en-masse, but this is something
people can make use of if they want to -and if we want to take advantage of it
across the test suite we can add tickets to address specific areas of concerns.