Fix currency validation bug
Before, when validating a currency that was `nil`, and the attribute was not
allowed to be blank, an error was raised.

This amends the `CurrencyCleaner` class so that `nil` values are cast to an
empty string, and the `spec/validators/currency_validator_spec.rb` is updated to
account for this.