Providers can specify client's cash income
Before, the cash income form relied on a legal aid application's transaction
types to determine which options to render in the form.

The enhanced bank upload journey does not create
`LegalAidApplicationTransactionType` records, but instead creates
`RegularTransaction` records. This meant that providers using the enhanced bank
upload journey were not presented with any options on the cash income page,
despite selecting regular payments in the previous step.

This adds a `cash_income_options` method to the `AggregatedCashIncome` model,
and updates the form to render inputs based on those options.

The method returns the correct transaction types depending on whether the
enhanced bank upload journey is enabled or not.

[Link to story](https://dsdmoj.atlassian.net/browse/AP-3467)