Introduce a `Flow::Steps::Step` struct
Each descendant of the `Flow::Flows::FlowSteps` class contains a mega `STEPS`
hash that spans hundreds of lines and each step often has associated complexity
and conditional logic.

These objects manage complex user flow and are an integral part of the
application's functionality, however they cannot be easily unit-tested. This
means we rely on more expensive integration tests to test each specific edge
case of flow logic.

This introduces a minimal `Flow::Steps::Step` struct, which can be used to
manage this complexity and introduces an interface for encapsulating this
important user flow logic that can be more easily unit-tested.