Use clamav in tests when explicitly tagged
Before, clamav was used for malware scanning in all tests that interacted with
the `MalwareScanner` class (either directly or indirectly).

While it is important and useful to test the entire stack in integration tests,
clamav is a source of repeated flakiness across the test suite due to the
latency it introduces (especially in CI).

This updates RSpec specs and Cucumber tests to mock a successful clamav response
unless they are explicitly tagged with `clamav: true` or `@clamav`.