Providers can submit housing benefit data
As part of the enhanced bank upload journey, providers need to tell us if their
client is in receipt of housing benefit or not. If they do receive housing
benefit, then they need to tell us how much, and how often.

If providers tell us that their client makes housing payments, then they are
asked whether the client receives housing benefit.

If they do, the information is used to create a regular transaction that can be
submitted to the CFE API to give an eligbility assessment result.

Housing benefits are credit transactions, and it makes sense for the application
to consider them as credit transactions.

There is a crucial distinction though between housing benefit and other credit
transaction types.

Disregarded benefits are benefits that do not count toward an applicant's
financial eligibility result, and housing benefits are a type of disregarded
benefit.

In most contexts, housing benefit transactions should be treated the same as
disregarded benefits -except when it comes to capturing regular transaction data
from providers.

In this context, we need to find out how much and how often applicants receive
housing benefit, so we can consider it alongside the housing payments they make
when calculating their financial eligibility result.

Similarly to `excluded_benefits`, `housing_benefit` is now defined as a child
transaction type of `benefits`. Excluded benefits encompasses all other types of
disregarded benefit types.

[Link to story](https://dsdmoj.atlassian.net/browse/AP-3462)